feat: split code used on server
